use double for Eig decomp

+++ include/patchwork_gpu/patchwork_gpu.cuh
@@ -234,11 +234,11 @@ class PatchWorkGPU {
   uint* patch_offsets_h{nullptr};  // For counting points in each patch
 
   PCAFeature* pca_features_d{nullptr};
-  float* cov_mats_d{nullptr};
+  double* cov_mats_d{nullptr};
   cusolverDnHandle_t cusolverH = NULL;
   syevjInfo_t syevj_params = NULL;
-  float* eigen_vals_d{nullptr};
-  float* work_d{nullptr};
+  double* eigen_vals_d{nullptr};
+  double* work_d{nullptr};
   int* eig_info_d{nullptr};
   std::vector<int> eig_info_h;
